Kami/requirements.txt
discord.py==2.4.0
python-dotenv==1.0.1
pymongo ==4.10.1
fastapi==0.07
